fix en.yml; fix method naming; rename number.rb to student.rb; refactor method 'telegram_api_response'
