+++ 3542/2/lib/telegram_file_downloader.rb
@@ -4,10 +4,10 @@ class TelegramFileDownloader
   API_URI_FOR_FILE_PATH = "https://api.telegram.org/bot#{ENV['BOT_TOKEN']}/getFile".freeze
   API_URI_FOR_DOWNLOADING = "https://api.telegram.org/file/bot#{ENV['BOT_TOKEN']}".freeze
 
-  attr_reader :file_identifier
+  attr_reader :file_id
 
-  def initialize(file_identifier)
-    @file_identifier = file_identifier
+  def initialize(file_id)
+    @file_id = file_id
   end
 
   def download_file
@@ -15,10 +15,10 @@ class TelegramFileDownloader
   end
 
   def image_uri
-    "#{API_URI_FOR_DOWNLOADING}/#{telegram_api_response['result']['file_path']}"
+    "#{API_URI_FOR_DOWNLOADING}/#{fetch_file_path}"
   end
 
-  def telegram_api_response
-    HTTParty.get(API_URI_FOR_FILE_PATH, query: { file_id: file_identifier }).to_h
+  def fetch_file_path
+    HTTParty.get(API_URI_FOR_FILE_PATH, query: { file_id: file_id }).to_h['result']['file_path']
   end
 end
