+++ 3542/2/validations/validations.rb
@@ -1,4 +1,3 @@
-require_relative 'number'
 require_relative 'registration.rb'
 require_relative 'selfie.rb'
 require_relative 'geo.rb'
